feat: integral support for storage transactions + bitcoin transaction processing in transaction (#1673)
## Description

Closes: #1194

## Changes

> [!NOTE]
Like 95% of this PR's line count is due to moving/splitting up our
stores (`in_memory.rs` and `postgres.rs`). I _did not_ alter any SQL, so
the fact that current integration tests also remain unchanged (short of
some minor tweaks in `use` statements etc.) and are passing should give
a fairly solid indication that the refactor generally does not _break_
anything. As a reviewer I would probably put most of my focus on
ensuring that no malicious code has snuck into the diffs.

- Restructured the `storage` module a bit:
  - Two new folders: `memory` and `pgsql`.
_(I had originally just named these `in_memory` and `postgres` but ran
into issues with the Emily container builds... 🤷 )_.
- Split `DbRead` and `DbWrite` implementations into `read.rs` and
`write.rs` for both memory/pgsql.
  - Core "store" logic is in `store.rs` for both memory/pgsql.
- New traits:
- `Transactable`: Implemented for stores which can create transactions
(i.e. our `SharedStore` and `PgStore`).
  - `TransactionHandle`: Implemented for transaction handle types
- New types:
- `InMemoryTransaction`: Represents a transaction for the in-memory
database and implements `TransactionHandle`. It uses a simple, naive
optimistic concurrency check for concurrent transactions.
- `PgTransaction`: Represents an sqlx postgres transaction and
implements `TransactionHandle`.
  - For postgres:
- New `PgRead` and `PgWrite` types which implement the "core" queries
using postgres `Executor`s, which the `DbRead` and `DbWrite`
implementations now delegate to.
- This adds a bit more boilerplate when adding new `Db*` methods, but
keeps the actual SQL logic centralized.
- I did try to find ways to minimize this, but so far haven't succeeded
to find another solution that works across-the-board like this, so I'm
ending my time-box and PR'ing the current solution.
- **Wraps bitcoin block+tx handling in a transaction** in
`block_observer.rs`.
- Updated tests & test helpers due to the above changes.

## Testing Information

## Checklist:

- [x] I have performed a self-review of my code
- [x] My changes generate no new warnings
- [x] New and existing unit tests pass locally with my changes
- [x] Any dependent changes have been merged and published in downstream
modules
