+++ signer/tests/integration/block_observer.rs
@@ -772,13 +772,7 @@ async fn block_observer_picks_up_chained_unordered_sweeps() {
         .with_first_bitcoin_core_client()
         .with_mocked_emily_client()
         .with_mocked_stacks_client()
-        .modify_settings(|settings| settings.signer.bootstrap_aggregate_key = Some(aggregate_key))
         .build();
-
-    let block_observer = BlockObserver {
-        context: ctx.clone(),
-        bitcoin_blocks: (),
-    };
 
     // Let's make sure the relevant tables are empty
     let sql = r#"
@@ -820,10 +814,14 @@ async fn block_observer_picks_up_chained_unordered_sweeps() {
     let mut transactions = block_info.transactions;
     transactions.shuffle(&mut rng);
 
-    block_observer
-        .extract_sbtc_transactions(block_hash, &transactions)
-        .await
-        .unwrap();
+    signer::block_observer::extract_sbtc_transactions(
+        &db,
+        Some(aggregate_key),
+        block_hash,
+        &transactions,
+    )
+    .await
+    .unwrap();
 
     // Okay, now the tables should be populated with all three
     // transactions.
