+++ supply-chain/config.toml
@@ -1421,10 +1421,6 @@ criteria = "safe-to-deploy"
 version = "0.2.3"
 criteria = "safe-to-deploy"
 
-[[exemptions.rustc_version]]
-version = "0.4.0"
-criteria = "safe-to-deploy"
-
 [[exemptions.rustfmt-wrapper]]
 version = "0.2.1"
 criteria = "safe-to-deploy"
